backend/requirements.txt
pillow
pyautogui
screeninfo